updated file headers, minor revisions
