+++ prompt_builder/augment_with_cfc.py
@@ -1,3 +1,16 @@
+# Copyright Amazon.com, Inc. or its affiliates. All rights reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import os
 import json
 import time
@@ -7,7 +20,7 @@ import multiprocessing as mp
 from tqdm import tqdm
 from functools import partial
 from rerank_utils import lexical_ranking, SemanticReranking
-from utils import file_distance, tokenize_nltk
+from utils import str2bool, file_distance, tokenize_nltk
 
 CHUNK_SIZE = 10
 SLIDING_WINDOW_SIZE = 10  # non-overlapping chunks if SLIDING_WINDOW_SIZE=CHUNK_SIZE
@@ -294,17 +307,6 @@ def attach_data(args, srcfile):
 
 
 if __name__ == "__main__":
-    def str2bool(v):
-        if isinstance(v, bool):
-            return v
-        if v.lower() in ('yes', 'true', 't', 'y', '1'):
-            return True
-        elif v.lower() in ('no', 'false', 'f', 'n', '0'):
-            return False
-        else:
-            raise argparse.ArgumentTypeError('Boolean value expected.')
-
-
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "--rerank",
